refactor core.py
